+++ NEWS.md
@@ -3,6 +3,8 @@
 * New `parse_openapi_spec()` and `parse_openapi_schema()` to convert an
   OpenAPI specification to a tibblify specification.
 
+* Fix ptype of a `tib_vector()` inside a `tib_df()`.
+
 * New `unpack_tspec()` to unpack the elements of `tib_row()` fields (#165).
 
 * Improved printing of lists parsed with `tspec_object()`.